requirements.txt
Django>=4.2.17,<4.3.0
djangorestframework>=3.15.2,<3.16
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.28.0,<0.29.0
django-profanity-filter>=0.2.1,<0.3.0
pytz==2024.2
djangorestframework-simplejwt>=5.4.0,<5.5.0